adjust: file utils and store pid in Long instead of string

+++ app/src/main/java/com/perol/asdpl/pixivez/activity/IntentActivity.kt
@@ -154,7 +154,7 @@ class IntentActivity : RinkActivity() {
 
     private fun openIllust(id: Long) {
         val bundle = Bundle()
-        val arrayList = longArrayOf(id.toLong())
+        val arrayList = longArrayOf(id)
         bundle.putLongArray("illustidlist", arrayList)
         bundle.putLong("illustid", id)
         val intent2 = Intent(this, PictureActivity::class.java)
